Added more things to the event system.
This time i completed most of the mouse_gui and now the player can walk
with pathfinding with the event system. With the event system i can even
make a post when the player have reached the destination of the path.
Making so that its easy to move and pick up stuff with only one click
from the user.